chore: move the version code from version.go into separate pkg
I missed that the build-strip Makefile step wrote a new version.go
file. Moving this code around is easier then messing with build
system right now.
