+++ internal/observability/metrics.go
@@ -11,6 +11,7 @@ import (
 	"github.com/sirupsen/logrus"
 	"github.com/supabase/auth/internal/conf"
 	"github.com/supabase/auth/internal/utilities"
+	"github.com/supabase/auth/internal/utilities/version"
 
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"go.opentelemetry.io/otel"
@@ -197,7 +198,7 @@ func ConfigureMetrics(ctx context.Context, mc *conf.MetricsConfig) error {
 			logrus.WithError(err).Error("unable to get gotrue.gotrue_running gague metric")
 			return
 		}
-		if err = utilities.InitVersionMetrics(ctx); err != nil {
+		if err = version.InitVersionMetrics(ctx, utilities.Version); err != nil {
 			logrus.WithError(err).Error("unable to configure version metrics")
 		}
 	})
